Continuando inserção de exceptions

+++ aba/src/main/java/com/aba/service/AlunoServiceImpl.java
@@ -3,12 +3,12 @@ import java.util.List;
 import java.util.Optional;
 
 import com.aba.dto.AlunosDTO;
-import com.aba.model.Instrutor;
-import com.aba.util.exception.CustomerException;
 import com.aba.interfaces.InstrutorService;
 import com.aba.interfaces.PlanoObjetivosService;
+import com.aba.model.Instrutor;
 import com.aba.model.PlanoObjetivos;
-import com.aba.util.exception.CustomerErro;
+import com.aba.util.CustomErrorType;
+import com.aba.util.MessageError;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -37,15 +37,12 @@ public class AlunoServiceImpl implements AlunoService {
     PlanoObjetivosService planoObjetivosService;
 
     public ResponseEntity<?> cadastrarAluno(AlunoDTO alunoDTO) {
-       Usuario aluno;
-       Instrutor instrutor;
-
-        try {
-            instrutor = this.instrutorService.getInstrutorByEmail(alunoDTO.getEmailInstrutor());
-        } catch (CustomerException e) {
-            return CustomerErro.instrutorInexistente(instrutor.getEmail());
+        Instrutor instrutor = this.instrutorService.getInstrutorByEmail(alunoDTO.getEmailInstrutor());
+        if(instrutor == null){
+            return MessageError.erroInstrutorNaoEncontradoByEmail(alunoDTO.getEmailInstrutor());
         }
 
+        Usuario aluno;
         aluno = new Aluno(alunoDTO.getNome(), alunoDTO.getIdade(),
                 instrutor, alunoDTO.getContato(), alunoDTO.getGenero(),
                 alunoDTO.getCpf(), alunoDTO.getEndereco(),
@@ -57,14 +54,12 @@ public class AlunoServiceImpl implements AlunoService {
         return ResponseEntity.status(HttpStatus.CREATED).body(((Aluno) aluno).getDtoCompleto());
     }
 
-    public ResponseEntity<?> editarAluno(Long id, AlunoDTO alunoDTO) {
-        Aluno aluno;
-
-        try {
-            aluno = this.getAlunoById(id);
-        } catch (CustomerException e) {
-            return CustomerErro.alunoInexistente(id);
+    public ResponseEntity<?> editarAluno(Long id, AlunoDTO alunoDTO){
+        Aluno aluno = this.getAlunoById(id);
+        if(aluno == null){
+            return MessageError.erroAlunoNaoEncontrado(id);
         }
+
         aluno.editar(alunoDTO, this.instrutorService.getInstrutorByEmail(alunoDTO.getEmailInstrutor()));
         this.alunoRepository.save(aluno);
 
@@ -72,10 +67,9 @@ public class AlunoServiceImpl implements AlunoService {
     }
 
     public ResponseEntity<?> removerAluno(Long id) {
-        try {
-            Aluno aluno = this.getAlunoById(id);
-        } catch (CustomerException e) {
-            return CustomerErro.alunoInexistente(id);
+        Aluno aluno = this.getAlunoById(id);
+        if(aluno == null){
+            return MessageError.erroAlunoNaoEncontrado(id);
         }
         this.usuarioRepository.deleteById(id);
 
@@ -90,16 +84,15 @@ public class AlunoServiceImpl implements AlunoService {
     }
 
     public ResponseEntity<?> consultarAluno(Long id) {
-        Aluno aluno;
-        try {
-            aluno = this.getAlunoById(id);
-        } catch (CustomerException e) {
-            return CustomerErro.alunoInexistente(id);
+        Aluno aluno = this.getAlunoById(id);
+        if(aluno == null){
+            return MessageError.erroAlunoNaoEncontrado(id);
         }
+
         return ResponseEntity.status(HttpStatus.OK).body(aluno.getDto());
     }
 
-    public ResponseEntity<?> adicionarPlanoObjetivos(Long id, Long idPlano) throws CustomerException {
+    public ResponseEntity<?> adicionarPlanoObjetivos(Long id, Long idPlano){
         Aluno aluno;
         aluno = getAlunoById(id);
         PlanoObjetivos planoObjetivos;
@@ -111,7 +104,7 @@ public class AlunoServiceImpl implements AlunoService {
         return ResponseEntity.status(HttpStatus.OK).body(aluno.getDtoCompleto());
     }
 
-    public ResponseEntity<?> removerPlanoObjetivos(Long id, Long idPlano) throws CustomerException {
+    public ResponseEntity<?> removerPlanoObjetivos(Long id, Long idPlano){
         Aluno aluno;
         aluno = getAlunoById(id);
         PlanoObjetivos planoObjetivos;
@@ -123,13 +116,12 @@ public class AlunoServiceImpl implements AlunoService {
         return ResponseEntity.status(HttpStatus.OK).body(aluno.getDtoCompleto());
     }
 
-    public Aluno getAlunoById(Long id) throws CustomerException {
+    public Aluno getAlunoById(Long id){
         Optional<Aluno> alunoOptional = this.alunoRepository.findById(id);
 
         if (!alunoOptional.isPresent()) {
-            throw new CustomerException("O aluno com id mencionado não existe");
-        }
-
-        return alunoOptional.get();
+            return null;
+        }return alunoOptional.get();
     }
+
 }
