+++ aba/src/main/java/com/aba/service/PlanoObjetivosServiceImpl.java
@@ -1,12 +1,13 @@
 package com.aba.service;
 
-import com.aba.dto.AtividadeDTO;
 import com.aba.dto.PlanoObjetivosDTO;
 import com.aba.dto.PlanosObjetivosDTO;
 import com.aba.interfaces.PlanoObjetivosService;
 import com.aba.model.Atividade;
+import com.aba.model.Instrutor;
 import com.aba.model.PlanoObjetivos;
 import com.aba.repository.PlanoObjetivosRepository;
+import com.aba.util.MessageError;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -28,17 +29,23 @@ public class PlanoObjetivosServiceImpl implements PlanoObjetivosService{
     InstrutorServiceImpl instrutorService;
 
     public ResponseEntity<?> cadastrarPlanoObjetivos(PlanoObjetivosDTO planoObjetivosDTO) {
-        PlanoObjetivos planoObjetivos;
-        planoObjetivos = new PlanoObjetivos(this.instrutorService.getInstrutorByEmail(planoObjetivosDTO.getEmailInstrutor()),
-                planoObjetivosDTO.getDescricaoAluno(), planoObjetivosDTO.getPrazo());
+        Instrutor instrutor = this.instrutorService.getInstrutorByEmail(planoObjetivosDTO.getEmailInstrutor());
+        if(instrutor == null){
+            return MessageError.erroInstrutorNaoEncontradoByEmail(planoObjetivosDTO.getEmailInstrutor());
+        }
+
+        PlanoObjetivos planoObjetivos = new PlanoObjetivos(instrutor, planoObjetivosDTO.getDescricaoAluno(),
+                                        planoObjetivosDTO.getPrazo());
 
         this.planoObjetivosRepository.save(planoObjetivos);
         return ResponseEntity.status(HttpStatus.OK).body(planoObjetivos.getDtoCompleto());
     }
 
     public ResponseEntity<?> editarPlanoObjetivos(Long id, PlanoObjetivosDTO planoObjetivosDTO) {
-        PlanoObjetivos planoObjetivos;
-        planoObjetivos = this.getPlanoObjetivosById(id);
+        PlanoObjetivos planoObjetivos = this.getPlanoObjetivosById(id);
+        if(planoObjetivos == null){
+            return MessageError.erroPlanoNaoEncontrado(id);
+        }
 
         planoObjetivos.editar(planoObjetivosDTO, instrutorService.getInstrutorByEmail(planoObjetivosDTO.getEmailInstrutor()));
         this.planoObjetivosRepository.save(planoObjetivos);
@@ -47,6 +54,10 @@ public class PlanoObjetivosServiceImpl implements PlanoObjetivosService{
     }
 
     public ResponseEntity<?> removerPlanoObjetivos(Long id) {
+        PlanoObjetivos planoObjetivos = this.getPlanoObjetivosById(id);
+        if(planoObjetivos == null){
+            return MessageError.erroPlanoNaoEncontrado(id);
+        }
         this.planoObjetivosRepository.deleteById(id);
 
         return ResponseEntity.status(HttpStatus.OK).body("Plano de Objetivos removido!");
@@ -54,6 +65,9 @@ public class PlanoObjetivosServiceImpl implements PlanoObjetivosService{
 
     public ResponseEntity<?> consultarPlanoObjetivos(Long id) {
         PlanoObjetivos planoObjetivos = this.getPlanoObjetivosById(id);
+        if(planoObjetivos == null){
+            return MessageError.erroPlanoNaoEncontrado(id);
+        }
 
         return ResponseEntity.status(HttpStatus.OK).body(planoObjetivos.getDtoCompleto());
     }
@@ -66,10 +80,15 @@ public class PlanoObjetivosServiceImpl implements PlanoObjetivosService{
     }
 
     public ResponseEntity<?> adicionarAtividadeNoPlano(Long idPlano, Long idAtividade) {
-        Atividade atividade;
-        atividade = this.atividadeService.getAtividadeById(idAtividade);
-        PlanoObjetivos planoObjetivos;
-        planoObjetivos = this.getPlanoObjetivosById(idPlano);
+        Atividade atividade = this.atividadeService.getAtividadeById(idAtividade);
+        if(atividade == null){
+            return MessageError.erroAtividadeNaoEncontrada(idAtividade);
+        }
+
+        PlanoObjetivos planoObjetivos = this.getPlanoObjetivosById(idPlano);
+        if(planoObjetivos == null){
+            return MessageError.erroPlanoNaoEncontrado(idPlano);
+        }
 
         planoObjetivos.addAtividade(atividade);
         this.planoObjetivosRepository.save(planoObjetivos);
@@ -78,10 +97,15 @@ public class PlanoObjetivosServiceImpl implements PlanoObjetivosService{
     }
 
     public ResponseEntity<?> removerAtividadeDoPlano(Long idPlano, Long idAtividade) {
-        Atividade atividade;
-        atividade = this.atividadeService.getAtividadeById(idAtividade);
-        PlanoObjetivos planoObjetivos;
-        planoObjetivos = this.getPlanoObjetivosById(idPlano);
+        Atividade atividade = this.atividadeService.getAtividadeById(idAtividade);
+        if(atividade == null){
+            return MessageError.erroAtividadeNaoEncontrada(idAtividade);
+        }
+
+        PlanoObjetivos planoObjetivos = this.getPlanoObjetivosById(idPlano);
+        if(planoObjetivos == null){
+            return MessageError.erroPlanoNaoEncontrado(idPlano);
+        }
 
         planoObjetivos.removeAtividade(atividade);
 
@@ -93,8 +117,6 @@ public class PlanoObjetivosServiceImpl implements PlanoObjetivosService{
 
         if (!planoObjetivos.isPresent()) {
             return null;
-        }
-
-        return planoObjetivos.get();
+        }return planoObjetivos.get();
     }
 }
